+++ lib/xlsxir.ex
@@ -2,7 +2,7 @@ defmodule Xlsxir do
   alias Xlsxir.{Unzip, SaxParser, Worksheet, Timer, Index}
 
   @moduledoc """
-  Extracts and parses data from a `.xlsx` file to an Erlang Term Storage (ETS) process and provides various functions for accessing the data.
+  Extracts and parses data from a `.xlsx` file to an Erlang Term Storage (ETS) process and provides various functions for accessing the data. **Warning:** empty cells are ignored.
   """
 
   @doc """
@@ -50,7 +50,54 @@ defmodule Xlsxir do
   end
 
   @doc """
-  Accesses `:worksheet` ETS process and returns data formatted as a list of row value lists.
+  Extracts worksheet data contained in the specified `.xlsx` file to an ETS process with a table identifier of `table_id` which is accessed via the `Xlsxir.Worksheet` module. Successful extraction 
+  returns `{:ok, table_id}` with the timer argument set to false and returns a tuple of `{:ok, table_id, time}` where `time` is a list containing time elapsed during the extraction process 
+  (i.e. `[hour, minute, second, microsecond]`) when the timer argument is set to true. The `table_id` is used to access data for that particular ETS process with the various access functions of the 
+  `Xlsxir` module.
+
+  Cells containing formulas in the worksheet are extracted as either a `string`, `integer` or `float` depending on the resulting value of the cell. 
+  Cells containing an ISO 8601 date format are extracted and converted to Erlang `:calendar.date()` format (i.e. `{year, month, day}`).
+
+  ## Parameters
+  - `path` - file path of a `.xlsx` file type in `string` format
+  - `index` - index of worksheet from within the Excel workbook to be parsed (zero-based index)
+  - `timer` - boolean flag that tracts extraction process time and returns it when set to `true`. Defalut value is `false`.
+
+  ## Example
+  Extract first worksheet in an example file named `test.xlsx` located in `./test/test_data`:
+
+        iex> {:ok, table_id} = Xlsxir.multi_extract("./test/test_data/test.xlsx", 0)
+        iex> table_id |> Xlsxir.Worksheet.alive?
+        true
+        iex> table_id |> Xlsxir.close
+        :ok
+  """
+  def multi_extract(path, index, timer \\ false) do
+    if timer, do: Timer.start
+
+    {:ok, file}       = Unzip.validate_path(path)
+    {:ok, file_paths} = Unzip.xml_file_list(index)
+                        |> Unzip.extract_xml_to_file(file)
+
+    Enum.each(file_paths, fn file -> 
+      case file do
+        'temp/xl/sharedStrings.xml' -> SaxParser.parse(to_string(file), :string)
+        'temp/xl/styles.xml'        -> SaxParser.parse(to_string(file), :style)
+        _                           -> nil
+      end
+    end)
+
+    {:ok, table_id} = SaxParser.parse("temp/xl/worksheets/sheet#{index + 1}.xml", :multi)
+    Unzip.delete_dir
+
+    if timer, do: {:ok, table_id, Timer.stop}, else: {:ok, table_id}
+  end
+
+  @doc """
+  Accesses ETS process and returns data formatted as a list of row value lists.
+
+  ## Parameters
+  - `table_id` - table identifier of ETS process to be accessed, defaults to `:worksheet`
 
   ## Example
   An example file named `test.xlsx` located in `./test/test_data` containing the following:
@@ -66,9 +113,15 @@ defmodule Xlsxir do
           [["string one", "string two", 10, 20, {2016, 1, 1}]]
           iex> Xlsxir.close
           :ok
+
+          iex> {:ok, table_id} = Xlsxir.multi_extract("./test/test_data/test.xlsx", 0)
+          iex> table_id |> Xlsxir.get_list
+          [["string one", "string two", 10, 20, {2016, 1, 1}]]
+          iex> table_id |> Xlsxir.close
+          :ok
   """
-  def get_list do
-    :ets.match(:worksheet, {:"$1", :"$2"})
+  def get_list(table_id \\ :worksheet) do
+    :ets.match(table_id, {:"$1", :"$2"})
     |> Enum.sort
     |> Enum.map(fn [_num, row] -> row
                                   |> Enum.map(fn [_ref, val] -> val end)
@@ -76,7 +129,10 @@ defmodule Xlsxir do
   end
 
   @doc """
-  Accesses `:worksheet` ETS process and returns data formatted as a map of cell references and values.
+  Accesses ETS process and returns data formatted as a map of cell references and values.
+
+  ## Parameters
+  - `table_id` - table identifier of ETS process to be accessed, defaults to `:worksheet`
 
   ## Example
   An example file named `test.xlsx` located in `./test/test_data` containing the following:
@@ -92,10 +148,15 @@ defmodule Xlsxir do
           %{ "A1" => "string one", "B1" => "string two", "C1" => 10, "D1" => 20, "E1" => {2016,1,1}}
           iex> Xlsxir.close
           :ok
+
+          iex> {:ok, table_id} = Xlsxir.multi_extract("./test/test_data/test.xlsx", 0)
+          iex> table_id |> Xlsxir.get_map
+          %{ "A1" => "string one", "B1" => "string two", "C1" => 10, "D1" => 20, "E1" => {2016,1,1}}
+          iex> table_id |> Xlsxir.close
+          :ok
   """
-  def get_map do
-    :ets.match(:worksheet, {:"$1", :"$2"})
-    |> Enum.sort
+  def get_map(table_id \\ :worksheet) do
+    :ets.match(table_id, {:"$1", :"$2"})
     |> Enum.reduce(%{}, fn [_num, row], acc -> 
          row
          |> Enum.reduce(%{}, fn [ref, val], acc2 -> Map.put(acc2, ref, val) end)
@@ -104,7 +165,10 @@ defmodule Xlsxir do
   end
 
   @doc """
-  Accesses `:worksheet` ETS process and returns an indexed map which functions like a multi-dimensional array in other languages.
+  Accesses ETS process and returns an indexed map which functions like a multi-dimensional array in other languages.
+
+  ## Parameters
+  - `table_id` - table identifier of ETS process to be accessed, defaults to `:worksheet`
 
   ## Example
   An example file named `test.xlsx` located in `./test/test_data` containing the following:
@@ -124,9 +188,19 @@ defmodule Xlsxir do
           10
           iex> Xlsxir.close
           :ok
+
+          iex> {:ok, table_id} = Xlsxir.multi_extract("./test/test_data/test.xlsx", 0)
+          iex> mda = table_id |> Xlsxir.get_mda
+          %{0 => %{0 => "string one", 1 => "string two", 2 => 10, 3 => 20, 4 => {2016,1,1}}}
+          iex> mda[0][0]
+          "string one"
+          iex> mda[0][2]
+          10
+          iex> table_id |> Xlsxir.close
+          :ok
   """
-  def get_mda do
-    :ets.match(:worksheet, {:"$1", :"$2"})
+  def get_mda(table_id \\ :worksheet) do
+    :ets.match(table_id, {:"$1", :"$2"})
     |> convert_to_indexed_map(%{})
   end
 
@@ -151,9 +225,10 @@ defmodule Xlsxir do
 
 
   @doc """
-  Accesses `:worksheet` ETS process and returns value of specified cell.
+  Accesses ETS process and returns value of specified cell.
 
   ## Parameters
+  - `table_id` - table identifier of ETS process to be accessed, defaults to `:worksheet`
   - `cell_ref` - Reference name of cell to be returned in `string` format (i.e. `"A1"`)
 
   ## Example
@@ -170,10 +245,23 @@ defmodule Xlsxir do
           "string one"
           iex> Xlsxir.close
           :ok
+
+          iex> {:ok, table_id} = Xlsxir.multi_extract("./test/test_data/test.xlsx", 0)
+          iex> table_id |> Xlsxir.get_cell("A1")
+          "string one"
+          iex> table_id |> Xlsxir.close
+          :ok
   """
-  def get_cell(cell_ref) do
+  def get_cell(cell_ref), do: do_get_cell(cell_ref)
+
+  @doc """
+  See `get_cell/1` documentation.
+  """
+  def get_cell(table_id, cell_ref), do: do_get_cell(cell_ref, table_id)
+
+  defp do_get_cell(cell_ref, table_id \\ :worksheet) do
     [[row_num]] = ~r/\d+/ |> Regex.scan(cell_ref)
-    [[row]]     = :ets.match(:worksheet, {row_num, :"$1"})
+    [[row]]     = :ets.match(table_id, {row_num, :"$1"})
 
     row
     |> Enum.filter(fn [ref, _val] -> ref == cell_ref end) 
@@ -185,6 +273,7 @@ defmodule Xlsxir do
   Accesses `:worksheet` ETS process and returns values of specified row in a `list`.
 
   ## Parameters
+  - `table_id` - table identifier of ETS process to be accessed, defaults to `:worksheet`
   - `row` - Reference name of row to be returned in `integer` format (i.e. `1`)
 
   ## Example
@@ -201,9 +290,22 @@ defmodule Xlsxir do
           ["string one", "string two", 10, 20, {2016, 1, 1}]
           iex> Xlsxir.close
           :ok
+
+          iex> {:ok, table_id} = Xlsxir.multi_extract("./test/test_data/test.xlsx", 0)
+          iex> table_id |> Xlsxir.get_row(1)
+          ["string one", "string two", 10, 20, {2016, 1, 1}]
+          iex> table_id |> Xlsxir.close
+          :ok
   """
-  def get_row(row) do
-    [[row]] = :ets.match(:worksheet, {to_string(row), :"$1"})
+  def get_row(row), do: do_get_row(row)
+
+  @doc """
+  See `get_row/1` documentation.
+  """
+  def get_row(table_id, row), do: do_get_row(row, table_id)
+
+  defp do_get_row(row, table_id \\ :worksheet) do
+    [[row]] = :ets.match(table_id, {to_string(row), :"$1"})
 
     row
     |> Enum.map(fn [_ref, val] -> val end)
@@ -213,6 +315,7 @@ defmodule Xlsxir do
   Accesses `:worksheet` ETS process and returns values of specified column in a `list`.
 
   ## Parameters
+  - `table_id` - table identifier of ETS process to be accessed, defaults to `:worksheet`
   - `col` - Reference name of column to be returned in `string` format (i.e. `"A"`)
 
   ## Example
@@ -229,9 +332,22 @@ defmodule Xlsxir do
           ["string one"]
           iex> Xlsxir.close
           :ok
+
+          iex> {:ok, table_id} = Xlsxir.multi_extract("./test/test_data/test.xlsx", 0)
+          iex> table_id |> Xlsxir.get_col("A")
+          ["string one"]
+          iex> table_id |> Xlsxir.close
+          :ok
   """
-  def get_col(col) do
-    :ets.match(:worksheet, {:"$1", :"$2"})
+  def get_col(col), do: do_get_col(col)
+
+  @doc """
+  See `get_col/1` documentation.
+  """
+  def get_col(table_id, col), do: do_get_col(col, table_id)
+
+  defp do_get_col(col, table_id \\ :worksheet) do
+    :ets.match(table_id, {:"$1", :"$2"})
     |> Enum.sort
     |> Enum.map(fn [_num, row] -> row
                                   |> Enum.filter_map(fn [ref, _val] -> 
@@ -243,11 +359,7 @@ defmodule Xlsxir do
   end
 
   @doc """
-  Returns count data based on `num_type` specified, default is `:all`. 
-  - `:rows` - Returns number of rows contained in worksheet
-  - `:cols` - Returns number of columns contained in worksheet
-  - `:cells` - Returns number of cells contained in worksheet
-  - `:all` - Returns a keyword list containing all of the above
+  See `get_multi_info\2` documentation.
   """
   def get_info(num_type \\ :all) do
     case num_type do
@@ -262,18 +374,42 @@ defmodule Xlsxir do
     end
   end
 
-  defp row_num do
-    :ets.info(:worksheet, :size)
+  @doc """
+  Returns count data based on `num_type` specified: 
+  - `:rows` - Returns number of rows contained in worksheet
+  - `:cols` - Returns number of columns contained in worksheet
+  - `:cells` - Returns number of cells contained in worksheet
+  - `:all` - Returns a keyword list containing all of the above
+  
+  ## Parameters
+  - `table_id` - table identifier of ETS process to be accessed, defaults to `:worksheet`
+  - `num_type` - type of count data to be returned (see above), defaults to `:all`
+  """
+  def get_multi_info(table_id, num_type \\ :all) do
+    case num_type do
+    :rows  -> row_num(table_id)
+    :cols  -> col_num(table_id)
+    :cells -> cell_num(table_id)
+    _      -> [
+                rows:  row_num(table_id),
+                cols:  col_num(table_id),
+                cells: cell_num(table_id)
+              ]
+    end
   end
 
-  defp col_num do
-    :ets.match(:worksheet, {:"$1", :"$2"})
-    |> Enum.scan(0, fn [_num, row], acc -> acc = Enum.count(row) end)
+  defp row_num(table_id \\ :worksheet) do
+    :ets.info(table_id, :size)
+  end
+
+  defp col_num(table_id \\ :worksheet) do
+    :ets.match(table_id, {:"$1", :"$2"})
+    |> Enum.map(fn [_num, row] -> Enum.count(row) end)
     |> Enum.max
   end
 
-  defp cell_num do
-    :ets.match(:worksheet, {:"$1", :"$2"})
+  defp cell_num(table_id \\ :worksheet) do
+    :ets.match(table_id, {:"$1", :"$2"})
     |> Enum.reduce(0, fn [_num, row], acc -> acc + Enum.count(row) end)
   end
 
@@ -287,9 +423,13 @@ defmodule Xlsxir do
       :ok
       iex> Xlsxir.close
       :ok
+
+      iex> {:ok, table_id} = Xlsxir.multi_extract("./test/test_data/test.xlsx", 0)
+      iex> Xlsxir.close(table_id)
+      :ok
   """
-  def close do
-    Worksheet.delete
+  def close(table_id \\ :worksheet) do
+    Worksheet.delete(table_id)
     |> case do
          false -> raise "Unable to close worksheet"
          true -> :ok